[expander] cross-phase persistent: allow more quote, quote-syntax
Previously, `quote` at the module level (i.e. on the rhs of a `define-values`
and not wrapped in a lambda, i.e. would run at module instantiation time) was
limited to a small subset of atomic data, and `quote-syntax` was forbidden to
appear anywhere in the module.

Loosen these restrictions, significantly. Specifically:

Define *simple data* as:

- numbers (including extflonums), booleans, symbols, chars, keywords, null
- immutable strings and immutable bytes
- immutable vectors with simple data for values
- immutable hasheqs with simple data for keys and values
- immutable boxes with simple data for content
- immutable prefab structs with simple data for field values

(These are the things the standard reader can read, except that it excludes
non-hasheq hashes, because cross-phase-persistent modules previously allowed
hasheqs but not other hashes.)

Then, the restrictions for `quote` are then relaxed so that any simple data can
be quoted at the module level. (As before, arbitrary data can be quoted when
guarded by a lambda.)

For `quote-syntax`, the restriction to simple data is enforced throughout the
module. This is necessary because the stx's quoted by a quote-syntax are, in
essence, lifted to the module level.