enable language-family-oriented documentation listings
Building on support for making searches specific to a language family,
enable documentation listings that are oriented to a language family.
All documentation is shown for any language choice, but the
organization of the listing can change based on the family.

In part, this an alternative to configuring an entire installation
around a language family other than "Racket". Different entry points
can be created for different language families into a single rendering
of documentation. To put it another way, we won't need a separate
pkg-build process to generate `docs.rhombus-lang.org` for Rhombus, and
the `docs.racket-lanf.org` build will support a Rhombus-orinteded (or
other-language-oriemted) view of the documentation.

Some aspects of documentation carry beyond the entry point, such as a
disposition of for searching or the main page to return to via the
"top" link. For that, we ressurrect and repair an idea for remembering
a user-specific entry point into the documentation, so it can be used
for later searches. That old facility broke when browsers stopped
allowing different `file://` URLs to share local storage. The
capability is restored here by putting the root path into the query
part of the URL, relying on even older support to thread query
parameters through documentation links. As a small refinement,
parameters that start with "q" are not propagated, which cleans up
some pesky search quesries that we don't want to be so sticky.

Building on that layer, a `fam` query is now recognized as providing a
language family to use as the default for searches. For example,
"fam=Rhombus" causes Rhombus search results to be shown first, and
Racket results are specially higlighted as such, instead of the other
way around. If a `famroot` query is also provided, then it names a
target document for "top" links; so, someone browing Rhombus
documentaion can easily get back to the Rhombus-oriented listing of
all documentation.

An example of putting everything together:

 * The Rhombus package can create a "rhombus" document where
   "rhombus/index.html" is a listing of all documentation, but from a
   Rhombus perspective: Rhombus documents are shown first, and RacKet
   documents are in categories that explicitly say "Racket".

 * A link to all Rhombus documentation could be

    file://<path>/rhombus/index.html?fam=Rhombus&famroot=rhombus

  or

    https://docs.racket-lang.org/rhombus/index.html?fam=Rhombus&famroot=rhombus

   which starts at the Rhombus documentation listing, causes searches
   to sort "Rhombus" family results before others, and causes the
   "top" links everywhere to return to "rhombus/index.html" instead of
   "index.html".

Although "info.rkt" files as of v8.14.0.5 support a language-family
specification with documentation, it's not flexible enough to support
all the variation that seems needed. So, another component of the
change here is to start pushing these kinds of properties into the
document's representation, instead of "info.rkt". Again, this builds
on the strategy started for indexing where the `tag-prefix` field of
`part` was hijacked for extension. In the long run, documentation
categories will be more conveniently specified in the ".scrbl" source
instead of in info.rkt" files.

Finally, this adds a "long key" notion to searching so that an index
entry can have two forms: one that has exactly the exported binding's
name, and one that includes, say, a conventional prefix. The latter is
common in Rhombus documentation, such as for `cmdline.parse`.

Some parts of this change require coodination through an update to the
"scribble" package.